fix: Specify client authn method
This is needed because if the method is not specified, the go client
will try to hit the token endpoint again, but with a different client
authn method if the first one fails. The problem is that the go client
does not try to parse the error it got from the provider, which in the
device flow most of the time will be authorization_pending.